chore: test `consistency` in `getDeployStore`
