+++ src/main.test.ts
@@ -39,7 +39,6 @@ const apiToken = 'some token'
 const signedURL = 'https://signed.url/123456789'
 const edgeToken = 'some other token'
 const edgeURL = 'https://edge.netlify'
-const uncachedEdgeURL = 'https://uncached.edge.netlify'
 
 describe('get', () => {
   describe('With API credentials', () => {
@@ -1491,185 +1490,6 @@ describe(`getStore`, () => {
 
     expect(() => getStore('production')).toThrowError(
       'Netlify Blobs could not find a `fetch` client in the global scope. You can either update your runtime to a version that includes `fetch` (like Node.js 18.0.0 or above), or you can supply your own implementation using the `fetch` property.',
-    )
-  })
-})
-
-describe('Consistency configuration', () => {
-  test('Respects the consistency mode supplied in the operation methods', async () => {
-    const mockMetadata = {
-      name: 'Netlify',
-      cool: true,
-      functions: ['edge', 'serverless'],
-    }
-    const headers = {
-      etag: '123456789',
-      'x-amz-meta-user': `b64;${base64Encode(mockMetadata)}`,
-    }
-    const mockStore = new MockFetch()
-      .get({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(value),
-        url: `${uncachedEdgeURL}/${siteID}/production/${key}`,
-      })
-      .head({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(null, { headers }),
-        url: `${uncachedEdgeURL}/${siteID}/production/${key}`,
-      })
-      .get({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(value, { headers }),
-        url: `${uncachedEdgeURL}/${siteID}/production/${key}`,
-      })
-
-    globalThis.fetch = mockStore.fetch
-
-    const context = {
-      edgeURL,
-      siteID,
-      token: edgeToken,
-      uncachedEdgeURL,
-    }
-
-    env.NETLIFY_BLOBS_CONTEXT = Buffer.from(JSON.stringify(context)).toString('base64')
-
-    const blobs = getStore('production')
-
-    const data = await blobs.get(key, { consistency: 'strong' })
-    expect(data).toBe(value)
-
-    const meta = await blobs.getMetadata(key, { consistency: 'strong' })
-    expect(meta?.etag).toBe(headers.etag)
-    expect(meta?.metadata).toEqual(mockMetadata)
-
-    const dataWithMeta = await blobs.getWithMetadata(key, { consistency: 'strong' })
-    expect(dataWithMeta?.data).toBe(value)
-    expect(dataWithMeta?.etag).toBe(headers.etag)
-    expect(dataWithMeta?.metadata).toEqual(mockMetadata)
-
-    expect(mockStore.fulfilled).toBeTruthy()
-  })
-
-  test('Respects the consistency mode supplied in the store constructor', async () => {
-    const mockMetadata = {
-      name: 'Netlify',
-      cool: true,
-      functions: ['edge', 'serverless'],
-    }
-    const headers = {
-      etag: '123456789',
-      'x-amz-meta-user': `b64;${base64Encode(mockMetadata)}`,
-    }
-    const mockStore = new MockFetch()
-      .get({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(value),
-        url: `${uncachedEdgeURL}/${siteID}/production/${key}`,
-      })
-      .head({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(null, { headers }),
-        url: `${uncachedEdgeURL}/${siteID}/production/${key}`,
-      })
-      .get({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(value, { headers }),
-        url: `${uncachedEdgeURL}/${siteID}/production/${key}`,
-      })
-
-    globalThis.fetch = mockStore.fetch
-
-    const blobs = getStore({
-      consistency: 'strong',
-      edgeURL,
-      name: 'production',
-      token: edgeToken,
-      siteID,
-      uncachedEdgeURL,
-    })
-
-    const data = await blobs.get(key)
-    expect(data).toBe(value)
-
-    const meta = await blobs.getMetadata(key)
-    expect(meta?.etag).toBe(headers.etag)
-    expect(meta?.metadata).toEqual(mockMetadata)
-
-    const dataWithMeta = await blobs.getWithMetadata(key)
-    expect(dataWithMeta?.data).toBe(value)
-    expect(dataWithMeta?.etag).toBe(headers.etag)
-    expect(dataWithMeta?.metadata).toEqual(mockMetadata)
-
-    expect(mockStore.fulfilled).toBeTruthy()
-  })
-
-  test('The consistency mode from the operation methods takes precedence over the store configuration', async () => {
-    const mockMetadata = {
-      name: 'Netlify',
-      cool: true,
-      functions: ['edge', 'serverless'],
-    }
-    const headers = {
-      etag: '123456789',
-      'x-amz-meta-user': `b64;${base64Encode(mockMetadata)}`,
-    }
-    const mockStore = new MockFetch()
-      .get({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(value),
-        url: `${uncachedEdgeURL}/${siteID}/production/${key}`,
-      })
-      .head({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(null, { headers }),
-        url: `${edgeURL}/${siteID}/production/${key}`,
-      })
-      .get({
-        headers: { authorization: `Bearer ${edgeToken}` },
-        response: new Response(value, { headers }),
-        url: `${edgeURL}/${siteID}/production/${key}`,
-      })
-
-    globalThis.fetch = mockStore.fetch
-
-    const blobs = getStore({
-      consistency: 'strong',
-      edgeURL,
-      name: 'production',
-      token: edgeToken,
-      siteID,
-      uncachedEdgeURL,
-    })
-
-    const data = await blobs.get(key)
-    expect(data).toBe(value)
-
-    const meta = await blobs.getMetadata(key, { consistency: 'eventual' })
-    expect(meta?.etag).toBe(headers.etag)
-    expect(meta?.metadata).toEqual(mockMetadata)
-
-    const dataWithMeta = await blobs.getWithMetadata(key, { consistency: 'eventual' })
-    expect(dataWithMeta?.data).toBe(value)
-    expect(dataWithMeta?.etag).toBe(headers.etag)
-    expect(dataWithMeta?.metadata).toEqual(mockMetadata)
-
-    expect(mockStore.fulfilled).toBeTruthy()
-  })
-
-  test('Throws when strong consistency is used and no `uncachedEdgeURL` property has been defined', async () => {
-    const context = {
-      edgeURL,
-      siteID,
-      token: edgeToken,
-    }
-
-    env.NETLIFY_BLOBS_CONTEXT = Buffer.from(JSON.stringify(context)).toString('base64')
-
-    const store = getStore('productin')
-
-    await expect(async () => await store.get('my-key', { consistency: 'strong' })).rejects.toThrowError(
-      "Netlify Blobs has failed to perform a read using strong consistency because the environment has not been configured with a 'uncachedEdgeURL' property",
     )
   })
 })
